+++ src/shared-components/chip/test.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import Tags from './index';
+import Chip from './index';
 
-describe('<Tags />', () => {
+describe('<Chip />', () => {
   describe('UI snapshots', () => {
     it('renders the correct css and text', () => {
       const component = renderer.create(
-        <Tags status="success" text="Success" />
+        <Chip status="success" text="Success" />
       );
 
       const tree = component.toJSON();
