+++ src/shared-components/index.js
@@ -1,4 +1,4 @@
 export { default as Alert } from './alert';
 export { default as RadioButton } from './radioButton';
-export { default as Tag } from './tag';
+export { default as Chip } from './chip';
 export { default as Typography, style as TYPOGRAPHY_STYLE } from './typography';
